requirements.txt
requests==2.25.1
python-whois==0.7.3
ipwhois==1.2.0
beautifulsoup4==4.9.3
google==3.0.0
googlesearch-python==2020.0.2
builtwith==1.3.4